refactor: windguru scraper
